+++ solaris_oci/cli/cli.py
@@ -15,12 +15,21 @@
 
 import argparse
 import importlib
-
+import logging
+from solaris_oci.version import __version__
 from solaris_oci.oci import oci_config
 from .container import Container
 from .volume import Volume
 from .image import Image
- 
+
+log_levels = {
+    'debug': logging.DEBUG, 
+    'info': logging.INFO, 
+    'warn': logging.WARNING, 
+    'error': logging.ERROR, 
+    'critical': logging.CRITICAL
+}
+
 class CustomFormatter(argparse.ArgumentDefaultsHelpFormatter, 
                       argparse.RawDescriptionHelpFormatter):
     pass        
@@ -36,11 +45,23 @@ class CLI:
         parser = argparse.ArgumentParser(
             formatter_class=CustomFormatter,
             description='A self-sufficient runtime for containers')
-        parser.add_argument('-D', '--debug',
-            help='Enable debug mode', 
-            action='store_true')
         parser.add_argument('-v', '--version',
             help='Print version information and quit', 
+            action='version',
+            version='%(prog)s version ' + __version__)
+        parser.add_argument('-l', '--log-level', 
+            help='Set the logging level ("debug"|"info"|"warn"|"error"|"fatal")',
+            choices=[
+                'debug',
+                'info',
+                'warn',
+                'error',
+                'critical'
+            ],
+            metavar='string',
+            default='info')
+        parser.add_argument('-D', '--debug',
+            help='Enable debug mode', 
             action='store_true')
         parser.add_argument('--root', 
             help='root directory for storage',
@@ -56,6 +77,8 @@ class CLI:
             command.init_parser(oci_subparsers)
  
         options = parser.parse_args()
+
+        logging.basicConfig(level=log_levels[options.log_level])
 
         if options.debug:
             import ptvsd