+++ solaris_oci/oci/graph/driver.py
@@ -12,5 +12,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-class Driver:
+# Singleton pattern as in
+# https://stackoverflow.com/questions/6760685/creating-a-singleton-in-python
+
+class Singleton(type):
+    _instances = {}
+    def __call__(cls, *args, **kwargs):
+        if cls not in cls._instances:
+            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
+        return cls._instances[cls]
+
+class Driver(metaclass=Singleton):
     pass