+++ solaris_oci/oci/image/distribution.py
@@ -14,9 +14,12 @@
 
 import json
 import pathlib
+import logging
 from opencontainers.distribution.v1 import RepositoryList
 from solaris_oci.oci import oci_config, OCIError
 from . import Repository, ImageInUseException
+
+log = logging.getLogger(__name__)
 
 def split_image_name(image_name):
     splitted_image_name = image_name.split(':')            
@@ -28,6 +31,7 @@ def split_image_name(image_name):
 
 class Distribution():
     def __init__(self):
+        log.debug('Creating instance of %s()' % type(self).__name__)
         self.repositories = None
         distribution_file_path = pathlib.Path(oci_config['global']['path'], 
             'distribution.json')
@@ -39,30 +43,39 @@ class Distribution():
     def load(self):
         distribution_path = pathlib.Path(oci_config['global']['path'])
         distribution_file_path = distribution_path.joinpath('distribution.json')
+        log.debug('Start loading distribution file (%s)', distribution_file_path)
+        if not distribution_file_path.is_file():
+            raise OCIError('Distribution file (%s) does not exist' % distribution_file_path)
         repository_list = RepositoryList.from_file(distribution_file_path)
         self.repositories = {}
         for repository_name in repository_list.get('Repositories'):
+            log.debug('Found repository (%s)', repository_name)
             repository = Repository(repository_name)
             self.repositories[repository_name] = repository
+        log.debug('Finish loading distribution file (%s)', distribution_file_path)
 
     def create(self):
         distribution_file_path = pathlib.Path(oci_config['global']['path'], 
             'distribution.json')
+        log.debug('Start creating distribution file (%s)', distribution_file_path)
         if distribution_file_path.is_file():
             raise OCIError('Distribution file (%s) already exists' % distribution_file_path)
         self.repositories = {}
         self.save()
+        log.debug('Finish creating distribution file (%s)', distribution_file_path)
 
     def save(self):
         distribution_path = pathlib.Path(oci_config['global']['path'])
+        distribution_file_path = distribution_path.joinpath('distribution.json')
+        log.debug('Start saving distribution file (%s)', distribution_file_path)
         if not distribution_path.is_dir():
             distribution_path.mkdir(parents=True)
         repository_list_json = {
             'repositories': list(self.repositories.keys())
         }
         repository_list = RepositoryList.from_json(repository_list_json)
-        distribution_file_path = distribution_path.joinpath('distribution.json')
         repository_list.save(distribution_file_path)
+        log.debug('Finish saving distribution file (%s)', distribution_file_path)
 
     def get_repository(self, repository_name):
         try:
@@ -70,14 +83,17 @@ class Distribution():
         except:
             raise OCIError('Repository (%s) does not exist' % repository_name)
 
-    def create_image(self, image_name, rootfs_tar_file, config_json):
+    def create_image(self, image_name, rootfs_tar_file, image_config):
+        log.debug('Start creating image (%s)', image_name)
         repository_name, tag = split_image_name(image_name)
         repository = self.repositories.get(repository_name, None)
         if repository is None:
             repository = Repository(repository_name)
             self.repositories[repository_name] = repository
             self.save()
-        return repository.create_image(tag, rootfs_tar_file, config_json)
+        image = repository.create_image(tag, rootfs_tar_file, image_config)
+        log.debug('Finish creating image (%s)', image_name)
+        return image
 
     def get_image(self, image_ref):        
         # image_ref, can either be:
@@ -97,6 +113,7 @@ class Distribution():
         raise OCIError('Image (%s) is unknown' % image_ref)
 
     def remove_image(self, image_name):
+        log.debug('Start removing image (%s)', image_name)
         image = self.get_image(image_name)
         repository_name = image.repository
         tag = image.tag
@@ -105,3 +122,4 @@ class Distribution():
         if repository.index is None:
             del self.repositories[repository_name]
             self.save()
+        log.debug('Finish removing image (%s)', image_name)