+++ solaris_oci/oci/image/layer.py
@@ -14,17 +14,21 @@
 
 import json
 import pathlib
+import logging
 from opencontainers.image.v1 import MediaTypeImageLayerNonDistributableGzip, \
     MediaTypeImageLayerNonDistributableZfsXZ, \
     MediaTypeImageLayerNonDistributableZfs, Descriptor
 from solaris_oci.oci import oci_config, OCIError
 from solaris_oci.util import digest_to_id, id_to_digest
 from solaris_oci.util.file import compress, sha256sum, rm
-from solaris_oci.oci.graph import graph_driver, NodeInUseException
+from solaris_oci.oci.graph import Graph, NodeInUseException
 from solaris_oci.oci.image import LayerInUseException
+
+log = logging.getLogger(__name__)
 
 class Layer:
     def __init__(self, diff_id=None, id=None, parent=None):
+        log.debug('Creating instance of %s(%s)' % (type(self).__name__, id or ''))
         self.diff_id = diff_id
         self.id = id
         self.parent = parent
@@ -36,13 +40,13 @@ class Layer:
     @property
     def size(self):
         if self.diff_id is not None:
-            return graph_driver.size(self.diff_id)
+            return Graph.driver().size(self.diff_id)
         return None
 
     @property
     def path(self):
         if self.diff_id is not None:
-            return graph_driver.path(self.diff_id)
+            return Graph.driver().path(self.diff_id)
         return None
 
     @property
@@ -54,21 +58,29 @@ class Layer:
     def create_diff(self):
         if self.diff_id is not None:
             raise('Layer (%s) already created' % self.diff_id)
-        self.diff_id = graph_driver.create(self.parent_diff_id)
+        self.diff_id = Graph.driver().create(self.parent_diff_id)
 
     def commit_diff(self, compressed=True):
-        graph_driver.commit(self.diff_id)
+        log.debug('Start committing diff (%s) of layer (%s)' % 
+            (self.diff_id, self.id))
+        Graph.driver().commit(self.diff_id)
         layers_path = pathlib.Path(oci_config['global']['path'], 'layers')
         if not layers_path.is_dir():
             layers_path.mkdir(parents=True)
         # TODO: Move to zfs_driver
         layer_file_path = layers_path.joinpath('layer.zfs')
-        graph_driver.save(self.diff_id, layer_file_path)
+        log.debug('Start saving diff (%s) of layer (%s) to file (%s)' % 
+            (self.diff_id, self.id, str(layer_file_path)))
+        Graph.driver().save(self.diff_id, layer_file_path)
+        log.debug('Finish saving diff (%s) of layer (%s) to file (%s)' % 
+            (self.diff_id, self.id, str(layer_file_path)))
         media_type=MediaTypeImageLayerNonDistributableZfs
         if compressed:
+            log.debug('Start compressing file (%s)' % str(layer_file_path))
             if compress(layer_file_path, method='xz', parallel=True) != 0:
                 raise OCIError('Could not compress layer file (%s)' 
                     % str(layer_file_path))
+            log.debug('Finish compressing file (%s)' % str(layer_file_path))
             media_type=MediaTypeImageLayerNonDistributableZfsXZ
             layer_file_path = layers_path.joinpath('layer.zfs.xz')
         # End TODO
@@ -82,16 +94,26 @@ class Layer:
             size=layer_path.stat().st_size,
             media_type=media_type,
         )
+        log.debug('Finish committing diff (%s) of layer (%s)' % 
+            (self.diff_id, self.id))
         return layer_descriptor
    
     def add_file_to_diff(self, file_path, destination_path=None):
         if self.diff_id is not None:
-            graph_driver.add_file(self.diff_id, file_path, destination_path)
+            Graph.driver().add_file(self.diff_id, file_path, destination_path)
+   
+    def add_tar_file_to_diff(self, tar_file, destination_path=None):
+        if self.diff_id is not None:
+            log.debug('Start adding tar file to diff (%s) of layer (%s)' % 
+                (self.diff_id, self.id))
+            Graph.driver().add_tar_file(self.diff_id, tar_file, destination_path)
+            log.debug('Finish adding tar file to diff (%s) of layer (%s)' % 
+                (self.diff_id, self.id))
 
     def remove_diff(self):
         if self.diff_id is not None:
             try:
-                graph_driver.remove(self.diff_id)
+                Graph.driver().remove(self.diff_id)
             except NodeInUseException:
                 raise LayerInUseException('Layer (%s) is in use, can not remove' % self.id)
             self.diff_id = None
@@ -105,4 +127,4 @@ class Layer:
             self.id = None   
 
     def is_parent(self):
-        return graph_driver.is_parent(self.diff_id)
+        return Graph.driver().is_parent(self.diff_id)